Fix the routing SSL logic (see #2323)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2171
| Docs PR or issue | -

This PR replaces #2262 and #2279.

## As discussed at the developer's meeting:

- Only switch to using absolute paths in Contao 5
- Fix the routing in Contao 4.10:
  - Checkbox not checked -> HTTP (not also HTTPS anymore!)
  - Checkbox checked -> HTTPS
- Use radio buttons instead of a checkbox

## Why radio buttons do not work

In order to remain BC, the option values need to be `''` and `'1'`, however, our `RadioButton` class interprets `''` as "not checked" and therefore never renders the HTTP option as checked. Of course, we could adjust the `Widget::optionChecked()` method, however, it would most likely have side effects. I therefore went with a select menu:

<img width="921" alt="" src="https://user-images.githubusercontent.com/1192057/93860169-f43f5b00-fcbe-11ea-98ec-41722da76e48.png">

## What the PR fixes

The UrlGenerator now prepends the `http://` protocol if the current website uses HTTPS and the target website does not. ✅ 

## What the PR still does not fix

If both the current and the target website use HTTPS, the UrlGenerator still generates a network path, therefore we still need #2279 to actually fix #2171. 🤷🏻‍♂️

Commits
-------

f7d3ecce Fix the routing SSL logic
57e232a1 Handle network paths when generating front end URLs
37dcf661 Revert switching the field positions