Fix picker for providers with insert tags (see #2353)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2346
| Docs PR or issue | -

This fixes the picker for picker providers that return insert tags (see #2346). It uses the picker callback to transform the DCA values back to the appropriate picker values and also checks whether it is actually necessary to load the previous values - which it only should be for the `checkbox` picker type.

I've tested this solution with:

* single `fileTree` and `pageTree` widgets
* `fileTree` and `pageTree` widgets that allow multiple selections
* pickers that allow the selection of files, articles, news, events, pages (e.g. the `url` field of `tl_content`)

Commits
-------

aca48537 fix picker for providers with insert tags (#2346)
e72aff2d only load previous values for checkbox field type
fd709af8 CS