Reduce check for inlined services (see #2339)
Description
-----------

Our current implementation calls `$container->getRemovedIds()` for every class before checking if its a singleton. This results in a lot of overhead, because `$container->getRemovedIds()` loads the removed IDs from file on every access. I assume that's to not memory-cache the IDs, but in our case it results in a ton of unnecessary calls.

The behavior of `System::import()` and `System::importStatic()` have slightly changed:
In debug mode, the existing behavior will remain the same. In production, we use `getInstance()` or try to create a new instance of the class before checking for removed IDs.

Commits
-------

3e3fdd59 Reduce check for inlined services
3fc699b1 Correctly cache the removed service IDs
37e7afe4 Merge branch '4.9' into bugfix/system-import
2f6c51e6 CS and tests