Ignore logout URL if no user is present (see #2388)
Description
-----------

If you access the URL of a logout page while not being logged in, the following error will occur:

```
InvalidArgumentException:
Unable to generate a logout url for an anonymous token.

  at vendor/symfony/security-http/Logout/LogoutUrlGenerator.php:144
  at Symfony\Component\Security\Http\Logout\LogoutUrlGenerator->getListener(null)
     (vendor/symfony/security-http/Logout/LogoutUrlGenerator.php:95)
  at Symfony\Component\Security\Http\Logout\LogoutUrlGenerator->generateLogoutUrl(null, 0)
     (vendor/symfony/security-http/Logout/LogoutUrlGenerator.php:76)
  at Symfony\Component\Security\Http\Logout\LogoutUrlGenerator->getLogoutUrl()
     (vendor/contao/contao/core-bundle/src/Resources/contao/pages/PageLogout.php:57)
  at Contao\PageLogout->getResponse(object(PageModel), true)
     (vendor/contao/contao/core-bundle/src/Resources/contao/controllers/FrontendIndex.php:339)
  at Contao\FrontendIndex->renderPage(object(PageModel))
     (vendor/symfony/http-kernel/HttpKernel.php:158)
  at Symfony\Component\HttpKernel\HttpKernel->handleRaw(object(Request), 1)
     (vendor/symfony/http-kernel/HttpKernel.php:80)
  at Symfony\Component\HttpKernel\HttpKernel->handle(object(Request), 1, true)
     (vendor/symfony/http-kernel/Kernel.php:201)
  at Symfony\Component\HttpKernel\Kernel->handle(object(Request))
     (web/index.php:31)
  at require('web/index.php')
     (web/app.php:4)
```

This can happen if the front end user happens to click twice on the logout link for example, causing two requests and the second click happens before the response of the first request arrives - thus the user will already be logged out for the second request.

This PR checks whether a user is actually logged in and if not simply redirects to the redirect target directly, without redirecting to the logout URL first.

One might say: "but Fritz, you obviously should protect the logout page 🙃 ". Well, I don't necessarily agree with that. In my case, the logout page isn't actually available anywhere in a regular navigation module, but rather is only used in either a custom navigation or a direct link, _within protected pages_. Thus there was no actual need to protect the logout page itself. Also, if you have defined a `401` page in your site structure (which you should) the 401 page would simply be shown in that case (or its redirect will be executed) - which typically means a login page will be shown. I would not expect the log**out** URL to show me a log**in**. I would expect the logout URL to always redirect to the defined redirect page, no matter what.

Commits
-------

d30d59bd ignore logout URL if no user is present
87cec53f add issue number
cb860633 remove comment
8141286d Update core-bundle/src/Resources/contao/pages/PageLogout.php

Co-authored-by: Leo Feyer <github@contao.org>
33a5724f use token check instead
e8f35cd1 code style
507e5e32 Move the comment to the correct position