Add image_container class to default figure template (see #2379)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

I figured we should probably add the `image_container` class to the core's default `figure.html.twig` template as well so that people can use it as a drop-in-replacement for the `image.html5` template. 

If someone should not want this class to be present in their installation, only a minor template adjustment is needed.

Commits
-------

cda83aaa add 'image_container' class to default figure.html.twig template to be inline with the image.html5 template