Fix file manager performance with large non-image files (see #2614)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

If you have one or more large non-image files within `files/`, the TTFB in the file manager of the back end will be significantly impacted with each large file since `DC_Folder::generateTree` always accesses `$objFile->width` and `$objFile->height` for each and every file. This in turn will execute `getimagesize`, even if the file is not actually an image, but for example a large MP4 file instead.

» [blackfire.io profiling](https://blackfire.io/profiles/f454b839-686b-41d2-b3cd-d9b93f53b0b7/graph)

This PR reduces the TTFB significantly by adding a `$this->isImage` check within `__get('imageSize')`.

For example, in my test setup, I had 6 MP4 files in a test folder within `files/`, each ~2.04 GiB in size. With this folder open in the file manager, the TTFB was between _50 to 60 **seconds**_.

With this change, the TTFB is cut down to only only _400 to 800 **ms**_ (note: my setup is fairly slow for some reason, but that's besides the point).

Also this makes me wonder if we should store the image size in the database as well, in order to improve the performance of the file manager even more?

Commits
-------

741252e4 add isImage check for imageSize
46745915 ignore the man behind the curtain
0d896f02 reverse order in condition
911f52d0 move condition
c2345419 remove getimagesize call altogether
c99527d0 fix imageViewSize types
76041197 Revert "fix imageViewSize types"

This reverts commit c99527d0ae874c34dcdd556f4e0d787a0b6a60ec.
679a094c use empty check
193455f2 Merge remote-tracking branch 'origin/4.9' into fix-large-file-manager-performance
cc4c5fa3 do not use empty() directly on __get
a495145d code style
bef402c7 fix ImageFactoryTest
53588947 move service retrieval out of try-catch
e1e5ada2 add comment