Remove the legacy encryption logic (see #2694)
Description
-----------

We've deprecated it a while ago but that doesn't really help us much because it will break as of PHP 7.2 and we're now at a minimum required version of 7.3 for 4.11.

The deprecation message and all the code that still resides in the class somewhat still implies that it works. Which is does absolutely not. So I prefer to remove it completely from the core.

Commits
-------

49ad27e8 Cleanup: Remove the encryption logic which doesn't work anymore as of PHP 7.2+
74734573 CS