Always sort root pages by language first (see #2831)
Description
-----------

The fix in #2819 was not correct. Luckily, merging the changes in Contao 4.11 showed the problem.

**Example case:**
 - Root page `de`
 - No `index` page in `de`
 - Root Page `en`
 - `index` page in `en`

The changes in #2819 would sort the english `index` page before any root pages. This means even if a user requests `de` as preferred language, she would not land on the german home page. This means the changes in #2819 need to be reverted.

While debugging the routing through `route:match` (thanks @dmolineus) I noticed the fallback page is not preferred if no language is given. Applying the fallback sorting fixes the initial issue in #2819 as well, most likely because bubblesort now correctly compares the necessary pages.

Commits
-------

4f2dc898 Always sort root pages by language first