Start the session to check for a user token (see #2829)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2810


Explicitly bypassing the `_wdt` route as suggested in https://github.com/contao/contao/issues/2810#issuecomment-786569685 is not the correct solution. The back end user is always available from the session, but we need to start the session if no other service has done that before the `TokenChecker`.

Commits
-------

bb9ef3ab Start the session to check for a user token