Show the values of unknown options (see #2956)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2943

Closes #2943

I decided to use the format `Unknown option: j_foo` bei the one the the brackets is often used (this hopefully helps a litte more to recognize the unknown option is an error like state).

I modified the label in the english `default.xlf`. This must also be done in the other languages. Can somebody help me to get into transifex to get the job done?

Commits
-------

7f3a6446 Show the values of unknown options ( closes #2943 )