Fix filtering for recurring events (see #2931)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

In 7670ee1e9fc5adf63a73ce662e57d73ab63138f9 `filter => true` was replaced with `search => true` for `tl_calendar_events.recurring`. However this field is a checkbox and you cannot search through such fields. I think this has been changed on accident - @bytehead or was there a specific reason? Currently you can neither search nor filter for events that have been set to _recurring_.

Keep in mind that re-enabling the filtering for _recurring_ will increase the amount of filtering options for Contao 4.11, making the filters take up two rows instead of one. But that's just how it is if we want to be able to keep filtering for the appropriate fields.

Commits
-------

757991a7 fix recurring filter
cb4e4dd6 do not show filter for noComments