Add the `rawPlainText()` and `rawHtmlToPlainText()` template helper methods (see #3110)
Description
-----------

Most of the following PRs for JSON-LD will likely require raw values in the template so I wanted to get this in first.
It makes the templates a lot more concise rather than using e.g. the long `Contao\StringUtil::revertInputEncoding()` calls.

I've also made sure they all start using `raw` so it's very clear that they output raw stuff. And I think the `revertInputEncoding` in the names is not required here as we're explicitly in the template context.

Commits
-------

da9574c8 Added support for raw quick access methods in Contao templates
2f641405 Remove raw() and updated description of other methods