Use schema.org JSON-LD in the news-bundle (see #3119)
Description
-----------

Using the new means for the news-bundle :)

Note for @m-vo: I've considered having some of the data in the `NewsModel` so it can be reused by Twig but there's not really much to be reused here, really. The links require the router, `hasText` queries the DB...so I left it all to the template.

Commits
-------

fcb1ac98 Switch to JSON-LD for news-bundle
3331d422 Update news-bundle/src/Resources/contao/modules/ModuleNews.php

Co-authored-by: Leo Feyer <github@contao.org>
3e697894 Update news-bundle/src/Resources/contao/modules/ModuleNews.php

Co-authored-by: Leo Feyer <github@contao.org>
8f9484b9 Add schema.org data for those entries that do not have a reader
a5a1b12d Allow to extend and remove JSON-LD data
9bc85429 Remove complexity again and leave it all to the template
4a1ee46b CS