Revert 'Lazy-load the `rootFallbackLanguage` property' (see #3166)
Description
-----------

This reverts #3067. The root fallback language model is already stored in the registry:

https://github.com/contao/contao/blob/99e04328a3b8eb06baec750aa8142ceb5804981e/core-bundle/src/Resources/contao/models/PageModel.php#L890-L899

Therefore, the changes from #3067 do reduce the number of calls to the `PageMode::findPublishedFallbackByHostname()` method, however, they do **not** reduce the number of DB queries.

Search the code for `contao.dns-fallback` to learn more about the special handling we already have.

Commits
-------

cdd05523 Revert 'Lazy-load the `rootFallbackLanguage` property'
94ce5ac8 Remove a redundant line

Co-authored-by: Martin Auswöger <martin@auswoeger.com>