Exclude the migrations from coverage (see #3177)
Description
-----------

We do not require writing unit tests for migrations, because unit tests that only check if the SQL queries are passed as arguments are pointless. If a migration is more complex than just checking if a table and column exist, we **can** still add unit tests (we also have two migration test classes) but thanks to this change, we no longer **have to** add pointless test classes.

Commits
-------

9f379b05 Exclude the migrations from coverage
c1ee7674 Exclude migration tests in all bundles