Deprecated the return of null response in fragments (see #3258)
Description
-----------

This should never have made it into the code base. When I suggested this in https://github.com/contao/contao/pull/622 I actually thought you don't need to write `return …` in the `getResponse()` method. But writing `return null` does not make sense at all…

The idea is that we can change the return value of the abstract method again in Contao 5

Commits
-------

77ecc45a Deprecated the return of null response in fragments
965cee11 Merge branch '4.12' into bugfix/empty-fragment-response