Always refresh the template hierarchy in the debug:contao-twig command (see #3270)
Description
-----------

We should always reset the loader cache + rebuild the template hierarchy when a user runs `debug:contao-twig`. 

The old implementation had a stupid issue where the reset part was run _after_ outputting results, but I could also create scenarios* where the cache wasn't up to date and you would need to reset it manually and _only_ on the command line. Removing the option and always resetting has a much better DX and is quick enough that you won't notice it.

*) e.g. running `contao-console cache:pool:clear cache.system` = clearing the cache pool, no warming

/cc @ausi

Commits
-------

1c4d4a0e always refresh loader in debug:contao-twig command