Remove reference to other table from column classname in DC_Table (see #3323)
Description
-----------

When I reference to another table in the list.label.fields of a dca, I get the request as column classname.

```
...
'list' => [
  'label' => [
    'fields' => [
      'send_to_customer',
      'member:tl_member.CONCAT(lastname,", ",firstname)',
      'event:tl_calendar_events.CONCAT(title, DATE_FORMAT(FROM_UNIXTIME(startTime), CHAR(32, 91, 37, 100, 46, 37, 109, 46, 37, 89, 93)))',
      'amount',
      'tstamp'
    ]
  ]
],
...
```

![image](https://user-images.githubusercontent.com/87128053/129586079-9884b1fe-7a7c-4553-ada3-878104e927e3.png)

The bugfix removes everything from the colon.

![image](https://user-images.githubusercontent.com/87128053/129588142-26caaa5f-c2f2-4412-b302-0f3623441fd6.png)

Commits
-------

cfbec44a Remove reference to other table from column classname in DC_Table

When you referenced to another table in the list.label.fields of a dca, you got the request as column classname. This fix is cutting off the classname based on variable $field from the colon.
903c4e3b Update core-bundle/src/Resources/contao/drivers/DC_Table.php

Co-authored-by: Fritz Michael Gschwantner <fmg@inspiredminds.at>
b8904a81 CS