Correctly explode foreign keys in the DCA (see #3321)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #3318
| Docs PR or issue | -

When defining a `foreignKey` for a database relation, or a field label in `list.label.fields` for your DCA, you are able to write the following syntax: `tl_member.CONCAT("firstname", " ", "lastname")` (or `member:tl_member.CONCAT("firstname", " ", "lastname")` for `list.label.fields`). This string is then split in two by using the first occurrence of `.` as the delimiter. However, some places of the `explode()` call were missing the third parameter, causing the string to be split into more than two parts, if there happen to be dots in other places. This PR fixes that by adding the necessary third parameter to the remaining `explode()` calls.

Commits
-------

77a60b4d fix foreignKey labels
72c9f688 fix relation in DcaExtractor
865ad228 also limit key and table definition to two parts
ebb636e4 Apply suggestions from code review

Co-authored-by: Martin Auswöger <martin@auswoeger.com>