Set memory limit for sub-processes when running contao-setup (see #3456)
Description
-----------

Running `php -dmemory_limit=<limit> contao-setup` won't make the sub processes run with a memory limit of `<limit>`. In fact this is true for all arguments because the binary cannot know what the php process was called with. 

We should IMHO at least handle the case for the memory limit, though. This PR does this by reading the value via `ini_get` and adding it to the call if not empty.

/cc @rabauss

Commits
-------

74bafa4c set memory limit for sub processes
a623be39 always set memory limit
1b9400a3 don't mock but ini_set memory limit in tests
1c454cd7 Reset the memory limit after the test
2528d0ed Merge pull request #6 from leofeyer/fix/memory-limit

Reset the memory limit after the test
276dc13e CS