Allow WebP to PNG/JPG conversion (see #3475)
Description
-----------

Currently you can only let Contao convert from JPG/PNG _to_ WEBP via database image size definitions - but not the other way around. The only way to implement seamless JPG/PNG to WEBP conversion _and vice versa_ is via the bundle configuration, e.g.:

```yaml
contao:
    image:
        sizes:
            example:
                width: 100
                formats:
                    jpg: [webp, jpg]
                    png: [webp, png]
                    webp: [webp, jpg]
```

With this PR the same will also be possible via the image size configuration in the database/backend:

<img src="https://user-images.githubusercontent.com/4970961/133432845-bbbc5519-2b48-420e-9d05-7e0ffafd8c30.png" width="332">

This allows you to seamlessly convert selected JPGs to WEBP and selected WEBP images to JPG (for the fallback) - which in turn means it will not matter whether your editors select a JPG, PNG or WEBP image in the backend. The output will always be

```html
<picture>
  <source srcset="foobar.webp" type="image/webp" …>
  <img src="foobar.jpg" …>
</picture>
```

for example.

As discussed with @ausi previously we decided to implement this as a bugfix.

_Note:_ this will allow you to select both _WEBP to JPG_ and _WEBP to PNG_ at the same time, though as far as I tested it does not matter. Contao will always pick _WEBP to JPG_ in such a case. @ausi should we implement some special handling or error message when both are selected? The bundle configuration would of course give you an error, if you tried to use two `webp` keys.

_Question:_ should we also allow _WEBP to GIF_ there?

Commits
-------

8f2972ea allow webp to png/jpg conversion
75987b68 Merge image size formats
0d39107b Coding style

Co-authored-by: Leo Feyer <github@contao.org>