Do not automatically focus input fields with autocomplete (see #3668)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #3651

The autofocus filter now skips inputs with autocomplete (besides `autocomplete="off"`). This prevents issues when autofocusing inputs with autocomplete plugins as described in #3651.

It is possible that the focused input is no longer the first visible input field (e.g. when editing a user account, the second input "Name" receives the focus), which might confuse some editors. However, I think the advantage of not having to deal with autocomplete plugins outweighs this.

Commits
-------

d7638d94 ignore inputs with autocomplete when autofocusing
33813213 CS
4add65aa do not skip empty autocomplete inputs