Set language of redirect page in global context (see #3647)
Description
-----------

Currently the language of a redirect page is not set in global context. This is especially a problem if you use language dependent insert tags in the redirect.

Affected Version: 4.9+

I'm not sure if we need to add more lines from the `prepare` method of `PageRegular`. The supposed lines are sufficient for my problem with the missing language context.

https://github.com/contao/contao/blob/2a19b186afd59495e9fbd5174d0deb40543c5079/core-bundle/src/Resources/contao/pages/PageRegular.php#L63-L69

Commits
-------

dab5842f Set language of redirect page in global context
0c07c675 Remove TL_KEYWORDS

Co-authored-by: Leo Feyer <github@contao.org>