Correctly store the referrer in the back end (see #3690)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes https://github.com/contao/contao/issues/2722
| Docs PR or issue | -


`$ref` is the new one so this has to be the new data.
I also had to adjust the existing unit tests because they did not make any sense at all as both, the existing referer ID as well as the new one were configured to `dummyTestRefererId`. This is not what's happening. It changes on every request.

Commits
-------

abeb9c83 Fixed referer stored incorrectly