Decorate the access decision manager (see #3611)
Description
-----------

Contao only works with the priority strategy and in the back end, we have to allow if all voters abstain whereas in the front end, it's the opposite.

This PR implements this logic by decorating the core access decision manager and delegating it to the Symfony one (default security settings) in case it's not a Contao request to provide full compatibility with apps that add Contao as a bundle.

Commits
-------

ef9798d3 Decorate access decision manager for correct voting strategies in Contao scope
c06e2568 CS
a88fa0f1 Make inner services configurable too
cfcd4861 Only use one ADM for both, the Contao FE and BE
5e839586 Bring back unit test
8cb84f2e Only decide on main request
36c0a39a Merge branch '4.x' into feature/correct-access-decision-manager
4fc0f50a CS
1c64524f Rename the service ID of the priority access decision manager