Make sure .public is a file and not a directory (see #3470)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #3465
| Docs PR or issue | -

If you happen to have a _directory_ called `.public` inside one of your directories in `files/`, Contao will think that it is already public, because it uses `file_exists` - which also returns `true` for directories. However, such a directory will then never actually be public (the symlink command filters for files only), nor can it be made public via the file manager

This PR changes the respective `file_exists` calls to `is_file`. It also adds new exceptions to the `Folder` class when trying to protect or unprotect a directory that contains a `.public` directory, so that the `.public` directory does not get accidentally deleted.

Commits
-------

fd30a905 check for .public file
3cc31896 use is_dir instead