Delete uploads if form did not validate (see #3734)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1185
| Docs PR or issue | -

This is an alternative, less invasive approach to #3600 in order to fix #1185. Instead of introducing a new mechanic for the legacy framework, this PR instead fixes the specific, original problem by making sure the `Form` removes any uploaded files, if the form itself did not validate. `Form` already purges `$_SESSION['FILES']` when the form did not validate here:

https://github.com/contao/contao/blob/ba5f1ba905f54918641267f2e90965d3d1153df0/core-bundle/src/Resources/contao/forms/Form.php#L262-L271

so the `Form` might as well also properly delete the uploads.

Commits
-------

1686bf33 delete uploads if form did not validate