Add transitive dependency on symfony/twig-bridge (see #3758)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #...
| Docs PR or issue | contao/docs#...

We use functions like `{% trans_default_domain %}` which are part of `symfony/twig-bridge`. As a consequence, we have a transitive dependency on this package. This PR adds the missing dependency.

I noticed this issue when (while being on 4.13@dev) I was updated on symfony/twig-bridge v6.0 which is currently incompatible with Contao (because of the usage of AnonymousToken, for example).

Commits
-------

3d44d43d Add transitive dependency to twig-bridge