Add a DCA option to collapse inactive checkbox groups (see #3674)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #3629
| Docs PR or issue | https://github.com/contao/docs/pull/888


If a checkbox widget contains multiple groups of options, it would be useful if the widget could be initialized with most of the groups collapsed (see #3629).

To selectively add this feature to some checkbox widgets, this draft adds a new DCA option `fields.eval.checkboxGroup` which itself can take up to three boolean options:
- `collapseInactive`: Collapse groups with no selected values
- `fallbackToFirst`: Open the first group of the widget if there are no selected values
- `overwriteSession`: Allow the other two options to change a previous manual user interaction with the widget's groups

The options can be combined in any way.

For the extensive access rights checkbox widgets in `tl_user` and `tl_user_group`, I'd suggest enabling all three options.

This way, only the first group of options will be visible when opening the edit view for the first time, which reduces the list of options while still presenting the editor an idea of the options available (instead of closing all groups without `fallbackToFirst: true`).

Whether the session should be overwritten is debatable. To me, it seems appropriate for the extensive option lists in `tl_user` and `tl_user_group` and ensures a consistent view for the editor (always showing groups with selected options, if there are any).

Initial view:

> <img width="362" alt="Bildschirmfoto 2021-11-14 um 20 56 13" src="https://user-images.githubusercontent.com/1193129/141696590-47ef3410-c394-418c-82bf-61a5bd199d8f.png">

View with selected options:

> <img width="370" alt="Bildschirmfoto 2021-11-14 um 20 56 30" src="https://user-images.githubusercontent.com/1193129/141696594-c07611ac-ca70-4c97-90fa-33e81a4a7ea4.png">


### Questions & todos:
- Is a nested option really necessary or is a simple flag enough (e.g. `collapseInactiveGroups`, grouping all three options together)?
- Are there any more checkbox widgets that should get initialized this way?
- Extend the DCA eval documentation.

Commits
-------

982f3ca5 add checkboxGroup options for CheckBox widget
4fe99c22 merge options into collapseUncheckedOptionGroups
9ec34c81 rename flag to collapseUncheckedGroups
cb96e62f Merge branch '4.x' into feature/checkbox-group-collapse-inactive